requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
attrs == 21.2.0
python_telegram_bot == 13.5
td_ameritrade_python_api == 0.3.5
